requirements.txt
python-fasthtml==0.12.25
google-genai==0.8.0
python-multipart==0.0.6
Pillow==10.1.0
PyPDF2==3.0.1
pdf2image==1.16.3
